requirements.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
backoff==2.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==14.0
croniter==2.0.1
dagster==1.5.13
dagster-graphql==1.5.13
dagster-pipes==1.5.13
dagster-webserver==1.5.13
dnspython==2.4.2
docopt==0.6.2
docstring-parser==0.15
exceptiongroup==1.2.0
fsspec==2023.12.2
gql==3.4.1
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-health-checking==1.60.0
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.2.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pendulum==2.1.2
pipreqs==0.4.13
pluggy==1.3.0
protobuf==4.21.12
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==0.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.24
starlette==0.34.0
tabulate==0.9.0
tomli==2.0.1
toposort==1.10
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
universal-pathlib==0.1.4
urllib3==1.26.18
uvicorn==0.25.0
uvloop==0.19.0
watchdog==3.0.0
watchfiles==0.21.0
websockets==12.0
yarg==0.1.9
yarl==1.9.4
